# microservices/legal-advogado/swagger.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
...
  "paths": {
...
    "/api/v1/tipos-atendimento": {
      "get": {
        "tags": ["Tipos de Atendimento"],
        "summary": "Lista todos os tipos de atendimento",
        "description": "Retorna uma lista de todos os tipos de atendimento cadastrados.",
        "operationId": "getAllTiposAtendimento",
        "responses": {
          "200": {
            "description": "Tipos de Atendimento listados com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListTipoAtendimentoResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListTipoAtendimentoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tipos-atendimento/{id}": {
      "get": {
        "tags": ["Tipos de Atendimento"],
        "summary": "Busca um tipo de atendimento pelo ID",
        "description": "Retorna os detalhes de um tipo de atendimento específico.",
        "operationId": "getTipoAtendimentoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do Tipo de Atendimento",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Tipo de Atendimento encontrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseTipoAtendimentoResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Tipo de Atendimento não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseTipoAtendimentoResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseTipoAtendimentoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/locais-atuacao": {
      "get": {
        "tags": ["Locais de Atuação"],
        "summary": "Lista todos os locais de atuação",
        "description": "Retorna uma lista de todos os locais de atuação cadastrados.",
        "operationId": "getAllLocaisAtuacao",
        "responses": {
          "200": {
            "description": "Locais de Atuação listados com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListLocalAtuacaoResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListLocalAtuacaoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/locais-atuacao/{id}": {
      "get": {
        "tags": ["Locais de Atuação"],
        "summary": "Busca um local de atuação pelo ID",
        "description": "Retorna os detalhes de um local de atuação específico.",
        "operationId": "getLocalAtuacaoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do Local de Atuação",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Local de Atuação encontrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseLocalAtuacaoResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Local de Atuação não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseLocalAtuacaoResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseLocalAtuacaoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/idiomas": {
      "get": {
        "tags": ["Idiomas"],
        "summary": "Lista todos os idiomas",
        "description": "Retorna uma lista de todos os idiomas cadastrados.",
        "operationId": "getAllIdiomas",
        "responses": {
          "200": {
            "description": "Idiomas listados com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListIdiomaResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListIdiomaResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/idiomas/{id}": {
      "get": {
        "tags": ["Idiomas"],
        "summary": "Busca um idioma pelo ID",
        "description": "Retorna os detalhes de um idioma específico.",
        "operationId": "getIdiomaById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do idioma",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Idioma encontrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseIdiomaResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Idioma não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseIdiomaResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseIdiomaResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IdiomaResponseDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "codigo": { "type": "string" },
          "nivel": { "type": "string" }
        }
      },
      "LocalAtuacaoResponseDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" }
        }
      },
...
      "TipoAtendimentoResponseDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" }
        }
      },
...
      "BaseResponseListTipoAtendimentoResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TipoAtendimentoResponseDTO"
            }
          },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
      "BaseResponseTipoAtendimentoResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": { "$ref": "#/components/schemas/TipoAtendimentoResponseDTO" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "BaseResponseListLocalAtuacaoResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LocalAtuacaoResponseDTO" }
          },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
      "BaseResponseLocalAtuacaoResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": { "$ref": "#/components/schemas/LocalAtuacaoResponseDTO" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
      "BaseResponseListIdiomaResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/IdiomaResponseDTO" }
          },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
      "BaseResponseIdiomaResponseDTO": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["SUCESSO", "ERRO", "AVISO"] },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "data": { "$ref": "#/components/schemas/IdiomaResponseDTO" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}